Move test and remove pointless duplicate tests
